+++ inst/resources/init-rprofile.R
@@ -1,3 +1,3 @@
-#### -- Packrat Autoloader (version 0.4.3-3) -- ####
+#### -- Packrat Autoloader (version 0.4.3-4) -- ####
 source("packrat/init.R")
 #### -- End Packrat Autoloader -- ####